backend/requirements.txt
Pillow==2.8.1
backports.ssl-match-hostname==3.4.0.2
certifi==2015.04.28
decorator==3.4.2
easydict==1.6
greenlet==0.4.6
imageio==1.2
motor==0.2
motorengine==0.8.11
moviepy==0.2.2.11
numpy==1.9.2
pymongo==2.7
pypng==0.0.17
python-dateutil==2.4.2
six==1.9.0
tornado==4.1
toro==0.8
tqdm==1.0
wsgiref==0.1.2
